requirements.txt
python-telegram-bot==20.6
opencv-contrib-python==4.9.0.80
numpy==1.26.4